+++ src/Symfony/Component/Messenger/Tests/Fixtures/DummyReceiver.php
@@ -1,0 +1,64 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Messenger\Tests\Fixtures;
+
+use Symfony\Component\Messenger\Envelope;
+use Symfony\Component\Messenger\Transport\Receiver\ReceiverInterface;
+
+class DummyReceiver implements ReceiverInterface
+{
+    private $deliveriesOfEnvelopes;
+    private $acknowledgedEnvelopes;
+    private $acknowledgeCount = 0;
+    private $rejectCount = 0;
+
+    /**
+     * @param Envelope[][] $deliveriesOfEnvelopes
+     */
+    public function __construct(array $deliveriesOfEnvelopes)
+    {
+        $this->deliveriesOfEnvelopes = $deliveriesOfEnvelopes;
+    }
+
+    public function get(): iterable
+    {
+        $val = array_shift($this->deliveriesOfEnvelopes);
+
+        return $val ?? [];
+    }
+
+    public function ack(Envelope $envelope): void
+    {
+        ++$this->acknowledgeCount;
+        $this->acknowledgedEnvelopes[] = $envelope;
+    }
+
+    public function reject(Envelope $envelope): void
+    {
+        ++$this->rejectCount;
+    }
+
+    public function getAcknowledgeCount(): int
+    {
+        return $this->acknowledgeCount;
+    }
+
+    public function getRejectCount(): int
+    {
+        return $this->rejectCount;
+    }
+
+    public function getAcknowledgedEnvelopes(): array
+    {
+        return $this->acknowledgedEnvelopes;
+    }
+}
