+++ src/Symfony/Component/Messenger/Tests/Fixtures/ResettableDummyReceiver.php
@@ -1,0 +1,29 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Messenger\Tests\Fixtures;
+
+use Symfony\Contracts\Service\ResetInterface;
+
+class ResettableDummyReceiver extends DummyReceiver implements ResetInterface
+{
+    private $hasBeenReset = false;
+
+    public function reset()
+    {
+        $this->hasBeenReset = true;
+    }
+
+    public function hasBeenReset(): bool
+    {
+        return $this->hasBeenReset;
+    }
+}
